chat/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
datasets==2.12.0
dill==0.3.6
dnspython==2.3.0
filelock==3.12.0
frozenlist==1.3.3
fsspec==2023.5.0
googleapis-common-protos==1.56.4
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.54.2
huggingface-hub==0.15.1
idna==3.4
loguru==0.7.0
lz4==4.3.2
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
numpy==1.24.3
openai==0.27.7
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
pinecone-client==2.2.1
platformdirs==3.5.1
protobuf==3.19.3
pyarrow==12.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
responses==0.18.0
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
xxhash==3.2.0
yarl==1.9.2
